Many changes outlined below.
Test file autogenerated, still empty however. All tests fail.

The ArgumentHandler is now the CommandHandler. This now has the task of
fully checking the inputted command: the command itself, the parameters
and the flags. If everything is valid, it sets it's isValid boolean to
True. If anything is invalid, False.

Only if isValid is True does the seize() method run the command. The
tricky part of all this is displaying the proper message to the user effeciently.

The checking functionality in CommandHandler has been divided up into
three methods which check the command, parameters and then the flags (in
that order).

I am not very happy with the solution as it is, hopefully I will be
making some major code cleaning commits soon.